move parameters into separate cpp
